+++ core/src/trap/best_effort.rs
@@ -18,14 +18,14 @@ use std::io::Write;
 use crate::Error;
 use crate::trap::Trap;
 
-/// A default trap that sends errors to standard error if possible.
+/// A best-effort trap that sends errors to standard error if possible.
 ///
 /// If standard error is not available, it does nothing.
 #[derive(Debug, Default)]
 #[non_exhaustive]
-pub struct DefaultTrap {}
+pub struct BestEffortTrap {}
 
-impl Trap for DefaultTrap {
+impl Trap for BestEffortTrap {
     fn trap(&self, err: &Error) {
         let _ = writeln!(io::stderr(), "{err}");
     }